Moved camp_config.hpp to solver_config.hpp.
