+++ test/unit/configure/process/CMakeLists.txt
@@ -16,12 +16,3 @@ create_standard_test(NAME ternary_chemical_activation_config SOURCES test_ternar
 create_standard_test(NAME troe_config SOURCES test_troe_config.cpp)
 create_standard_test(NAME tunneling_config SOURCES test_tunneling_config.cpp)
 
-create_standard_test(NAME arrhenius_camp_config SOURCES test_arrhenius_camp_config.cpp)
-create_standard_test(NAME branched_camp_config SOURCES test_branched_camp_config.cpp)
-create_standard_test(NAME emission_camp_config SOURCES test_emission_camp_config.cpp)
-create_standard_test(NAME first_order_loss_camp_config SOURCES test_first_order_loss_camp_config.cpp)
-create_standard_test(NAME photolysis_camp_config SOURCES test_photolysis_camp_config.cpp)
-create_standard_test(NAME surface_camp_config SOURCES test_surface_camp_config.cpp)
-create_standard_test(NAME ternary_chemical_activation_camp_config SOURCES test_ternary_chemical_activation_camp_config.cpp)
-create_standard_test(NAME troe_camp_config SOURCES test_troe_camp_config.cpp)
-create_standard_test(NAME tunneling_camp_config SOURCES test_tunneling_camp_config.cpp)
