--- test/unit/configure/process/test_photolysis_camp_config.cpp
@@ -1,82 +1,0 @@
-#include <gtest/gtest.h>
-
-#include <micm/configure/camp_config.hpp>
-
-TEST(PhotolysisConfig, DetectsInvalidConfig)
-{
-  micm::SolverConfig solver_config;
-
-  // Read and parse the configure files
-  micm::ConfigParseStatus status = solver_config.ReadAndParse("./unit_configs/process/photolysis/missing_reactants");
-  EXPECT_EQ(micm::ConfigParseStatus::RequiredKeyNotFound, status);
-  status = solver_config.ReadAndParse("./unit_configs/process/photolysis/missing_products");
-  EXPECT_EQ(micm::ConfigParseStatus::RequiredKeyNotFound, status);
-  status = solver_config.ReadAndParse("./unit_configs/process/photolysis/missing_MUSICA_name");
-  EXPECT_EQ(micm::ConfigParseStatus::RequiredKeyNotFound, status);
-}
-
-TEST(PhotolysisConfig, ParseConfig)
-{
-  micm::SolverConfig solver_config;
-
-  micm::ConfigParseStatus status = solver_config.ReadAndParse("./unit_configs/process/photolysis/valid");
-  EXPECT_EQ(micm::ConfigParseStatus::Success, status);
-
-  micm::SolverParameters solver_params = solver_config.GetSolverParams();
-
-  auto& process_vector = solver_params.processes_;
-
-  // first reaction
-  {
-    EXPECT_EQ(process_vector[0].reactants_.size(), 1);
-    EXPECT_EQ(process_vector[0].reactants_[0].name_, "foo");
-    EXPECT_EQ(process_vector[0].products_.size(), 2);
-    EXPECT_EQ(process_vector[0].products_[0].first.name_, "bar");
-    EXPECT_EQ(process_vector[0].products_[0].second, 1.0);
-    EXPECT_EQ(process_vector[0].products_[1].first.name_, "baz");
-    EXPECT_EQ(process_vector[0].products_[1].second, 3.2);
-    micm::UserDefinedRateConstant* photo_rate_constant =
-        dynamic_cast<micm::UserDefinedRateConstant*>(process_vector[0].rate_constant_.get());
-    EXPECT_EQ(photo_rate_constant->SizeCustomParameters(), 1);
-    EXPECT_EQ(photo_rate_constant->CustomParameters()[0], "PHOTO.jfoo");
-  }
-
-  // second reaction
-  {
-    EXPECT_EQ(process_vector[1].reactants_.size(), 1);
-    EXPECT_EQ(process_vector[1].reactants_[0].name_, "bar");
-    EXPECT_EQ(process_vector[1].products_.size(), 2);
-    EXPECT_EQ(process_vector[1].products_[0].first.name_, "bar");
-    EXPECT_EQ(process_vector[1].products_[0].second, 0.5);
-    EXPECT_EQ(process_vector[1].products_[1].first.name_, "foo");
-    EXPECT_EQ(process_vector[1].products_[1].second, 1.0);
-    micm::UserDefinedRateConstant* photo_rate_constant =
-        dynamic_cast<micm::UserDefinedRateConstant*>(process_vector[1].rate_constant_.get());
-    EXPECT_EQ(photo_rate_constant->SizeCustomParameters(), 1);
-    EXPECT_EQ(photo_rate_constant->CustomParameters()[0], "PHOTO.jbar");
-  }
-}
-
-TEST(PhotolysisConfig, DetectsNonstandardKeys)
-{
-  micm::SolverConfig solver_config;
-
-  micm::ConfigParseStatus status = solver_config.ReadAndParse("./unit_configs/process/photolysis/contains_nonstandard_key");
-  EXPECT_EQ(micm::ConfigParseStatus::ContainsNonStandardKey, status);
-}
-
-TEST(PhotolysisConfig, DetectsNonstandardProductCoefficient)
-{
-  micm::SolverConfig solver_config;
-
-  micm::ConfigParseStatus status = solver_config.ReadAndParse("./unit_configs/process/photolysis/nonstandard_product_coef");
-  EXPECT_EQ(micm::ConfigParseStatus::ContainsNonStandardKey, status);
-}
-
-TEST(PhotolysisConfig, DetectsNonstandardReactantCoefficient)
-{
-  micm::SolverConfig solver_config;
-
-  micm::ConfigParseStatus status = solver_config.ReadAndParse("./unit_configs/process/photolysis/nonstandard_reactant_coef");
-  EXPECT_EQ(micm::ConfigParseStatus::ContainsNonStandardKey, status);
-}
